requirements.txt
pytesseract
Pillow
googletrans==4.0.0rc1
requests
